bugzilla 2029 - try again

+++ src/mars.h
@@ -364,7 +364,7 @@ enum MATCH
     MATCHexact		// exact match
 };
 
-typedef unsigned StorageClass;
+typedef uint64_t StorageClass;
 
 
 void warning(Loc loc, const char *format, ...);
